src/frontend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
cachetools==4.2.4
    # via google-auth
certifi==2021.10.8
    # via requests
cffi==1.14.6
    # via cryptography
charset-normalizer==2.0.6
    # via requests
click==8.0.2
    # via flask
colorama==0.4.4
    # via click
cryptography==3.3.2
    # via -r requirements.in
flask==1.1.2
    # via
    #   -r requirements.in
    #   opentelemetry-instrumentation-flask
google-api-core[grpc]==1.31.3
    # via
    #   google-cloud-core
    #   google-cloud-monitoring
    #   google-cloud-trace
google-auth==1.35.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   opentelemetry-tools-google-cloud
google-cloud-core==1.7.2
    # via google-cloud-trace
google-cloud-monitoring==1.1.0
    # via opentelemetry-exporter-google-cloud
google-cloud-trace==0.24.0
    # via opentelemetry-exporter-google-cloud
googleapis-common-protos==1.53.0
    # via google-api-core
greenlet==1.1.2
    # via sqlalchemy
grpcio==1.41.0
    # via google-api-core
gunicorn==20.1.0
    # via -r requirements.in
idna==3.2
    # via requests
itsdangerous==2.0.1
    # via flask
jinja2==2.11.3
    # via
    #   flask
    #   opentelemetry-instrumentation-jinja2
markupsafe==2.0.1
    # via jinja2
opentelemetry-api==0.13b0
    # via
    #   opentelemetry-exporter-google-cloud
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-jinja2
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-sqlalchemy
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
    #   opentelemetry-tools-google-cloud
opentelemetry-exporter-google-cloud==0.13b0
    # via -r requirements.in
opentelemetry-instrumentation==0.13b0
    # via
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-jinja2
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-sqlalchemy
    #   opentelemetry-instrumentation-wsgi
opentelemetry-instrumentation-flask==0.13b0
    # via -r requirements.in
opentelemetry-instrumentation-jinja2==0.13b0
    # via -r requirements.in
opentelemetry-instrumentation-requests==0.13b0
    # via -r requirements.in
opentelemetry-instrumentation-sqlalchemy==0.13b0
    # via -r requirements.in
opentelemetry-instrumentation-wsgi==0.13b0
    # via opentelemetry-instrumentation-flask
opentelemetry-sdk==0.13b0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-google-cloud
    #   opentelemetry-tools-google-cloud
opentelemetry-tools-google-cloud==0.13b0
    # via -r requirements.in
packaging==21.0
    # via google-api-core
protobuf==3.17.3
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.20
    # via cffi
pyjwt==1.7.1
    # via -r requirements.in
pyparsing==2.4.7
    # via packaging
pytz==2021.3
    # via google-api-core
requests==2.26.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-tools-google-cloud
rsa==4.7.2
    # via google-auth
six==1.16.0
    # via
    #   cryptography
    #   google-api-core
    #   google-auth
    #   google-cloud-core
    #   grpcio
    #   protobuf
sqlalchemy==1.4.25
    # via opentelemetry-instrumentation-sqlalchemy
urllib3==1.26.6
    # via
    #   -r requirements.in
    #   requests
werkzeug==2.0.2
    # via flask
wrapt==1.13.1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-jinja2
    #   opentelemetry-instrumentation-sqlalchemy

# The following packages are considered to be unsafe in a requirements file:
# setuptools
